Projects/youtubeNotes/requirements.txt
youtube-transcript-api
streamlit
openai
python-dotenv
pathlib
google-generativeai
langchain